Move RawLogDTO
