+++ src/main/java/qtlog/LogParser/ILogParser.java
@@ -1,7 +1,5 @@
 package qtlog.LogParser;
 
-import qtlog.shared.RawLogDTO;
-
 import java.nio.file.Path;
 
 public interface ILogParser {
